/trunk/core: - Fix character saving to DB - Remove unnecessary casts - ExShowDominionRegistry full structure - ExReplyRegisterDominion base structure
git-svn-id: http://svn.l2jfree.com/l2jfree@7463 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/actor/instance/L2PcInstance.java
@@ -7123,7 +7123,7 @@ public final class L2PcInstance extends L2Playable
 			statement.setString(48, getName());
 			statement.setLong(49, getDeathPenaltyBuffLevel());
 			statement.setLong(50, getTrustLevel());
-			statement.setDouble(51, getVitalityPoints());
+			statement.setInt(51, getVitalityPoints());
 			statement.setInt(52, getBookMarkSlot());
 		    statement.setInt(53, getObjectId());
 			statement.execute();
